+++ graylog2-web-interface/src/preflight/App.test.tsx
@@ -15,10 +15,13 @@
  * <http://www.mongodb.com/licensing/server-side-public-license>.
  */
 import React from 'react';
-import { screen, renderPreflight } from 'wrappedTestingLibrary';
+import { screen, renderPreflight, waitFor } from 'wrappedTestingLibrary';
 import userEvent from '@testing-library/user-event';
 
 import fetch from 'logic/rest/FetchProvider';
+import useDataNodes from 'preflight/hooks/useDataNodes';
+import { asMock } from 'helpers/mocking';
+import UserNotification from 'preflight/util/UserNotification';
 
 import App from './App';
 
@@ -38,6 +41,11 @@ jest.mock('preflight/hooks/useDataNodes', () => jest.fn(() => ({
   isInitialLoading: false,
   error: undefined,
 })));
+
+jest.mock('preflight/util/UserNotification', () => ({
+  error: jest.fn(),
+  success: jest.fn(),
+}));
 
 describe('App', () => {
   let windowConfirm;
@@ -64,13 +72,51 @@ describe('App', () => {
     await screen.findByText(/It looks like you are starting Graylog for the first time and have not configured a data node./);
   });
 
-  it('should render loading page after resuming startup', async () => {
+  it('should resume startup and display loading page', async () => {
     renderPreflight(<App />);
 
-    userEvent.click(await screen.findByRole('button', { name: /resume startup/i }));
+    const resumeStartupButton = await screen.findByRole('button', {
+      name: /resume startup/i,
+    });
 
+    userEvent.click(resumeStartupButton);
+
+    await waitFor(() => expect(fetch).toHaveBeenCalledWith('POST', expect.stringContaining('/api/status/finish-config'), undefined, false));
     await screen.findByText(/The Graylog server is currently starting./);
+  });
 
-    expect(fetch).toHaveBeenCalledWith('POST', expect.stringContaining('/api/status/finish-config'), undefined, false);
+  it('should display confirm dialog on resume startup when there is no Graylog data node', async () => {
+    asMock(useDataNodes).mockReturnValue({
+      data: [],
+      isFetching: false,
+      isInitialLoading: false,
+      error: undefined,
+    });
+
+    renderPreflight(<App />);
+
+    const resumeStartupButton = screen.getByRole('button', {
+      name: /resume startup/i,
+    });
+
+    userEvent.click(resumeStartupButton);
+
+    await waitFor(() => expect(window.confirm).toHaveBeenCalledWith('Are you sure you want to resume startup without a running Graylog data node?'));
+    await waitFor(() => expect(fetch).toHaveBeenCalledWith('POST', expect.stringContaining('/api/status/finish-config'), undefined, false));
+  });
+
+  it('should display error when resuming startup failed', async () => {
+    asMock(fetch).mockImplementation(() => Promise.reject(new Error('Unexpected error!')));
+
+    renderPreflight(<App />);
+
+    const resumeStartupButton = screen.getByRole('button', {
+      name: /resume startup/i,
+    });
+
+    userEvent.click(resumeStartupButton);
+
+    await waitFor(() => expect(fetch).toHaveBeenCalledWith('POST', expect.stringContaining('/api/status/finish-config'), undefined, false));
+    await waitFor(() => expect(UserNotification.error).toHaveBeenCalledWith('Resuming startup failed with error: Error: Unexpected error!', 'Could not resume startup'));
   });
 });
